Get Steal to tree-shake without transpile hook
The transpile hook is skipped during the build phase, the Babel plugin
that takes care of removing unused exports was run in this hook causing
steal-tools build to break on packages that are not being used.

This change makes it so that when the transpile hook is not called,
steal still runs the babel plugin to remove unused exports in the
`instantiate` hook attached by the trace extension.